// Client/src/app/core/modules/openapi/model/update-degree-program-specialization-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateDegreeProgramSpecializationDTO { 
    name?: string;
    responsibleUserId?: string;
}